fakechroot: Add the environment arch-chroot.env
The script pacstrap calls the mount, umount and unshare. The install
scripts of packages calls ldconfig, ldd.

Additionally, it check for the root privileges using the bash environ
variable EUID.

This adds the fakechroot environment pacstrap.env that replaces the
calls to mount, umount and ldconfig by true, and the calls to ldd and
unshare by the fakechroot wrappers. Also, it overwrites the EUID
variable to 0 to fake the user to root.